restructured reward card and jobbahut page - got reward card collection to work correctly
